+++ panther_utils/CMakeLists.txt
@@ -57,8 +57,7 @@ if(BUILD_TESTING)
     ${PROJECT_NAME}_test_ros_utils
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
            $<INSTALL_INTERFACE:include>)
-  ament_target_dependencies(${PROJECT_NAME}_test_ros_utils std_msgs tf2_ros
-                            tf2_geometry_msgs)
+  ament_target_dependencies(${PROJECT_NAME}_test_ros_utils std_msgs)
 
   ament_add_gtest(${PROJECT_NAME}_test_ros_test_utils
                   test/test_ros_test_utils.cpp)
@@ -67,6 +66,14 @@ if(BUILD_TESTING)
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
            $<INSTALL_INTERFACE:include>)
   ament_target_dependencies(${PROJECT_NAME}_test_ros_test_utils rclcpp std_msgs)
+
+  ament_add_gtest(${PROJECT_NAME}_test_tf2_utils test/test_tf2_utils.cpp)
+  target_include_directories(
+    ${PROJECT_NAME}_test_tf2_utils
+    PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+           $<INSTALL_INTERFACE:include>)
+  ament_target_dependencies(${PROJECT_NAME}_test_tf2_utils std_msgs tf2_ros
+                            tf2_geometry_msgs)
 
   ament_add_gtest(${PROJECT_NAME}_test_yaml_utils test/test_yaml_utils.cpp)
   target_include_directories(
