+++ panther_utils/include/panther_utils/ros_utils.hpp
@@ -12,14 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef PANTHER_UTILS__ROS_UTILS_HPP_
-#define PANTHER_UTILS__ROS_UTILS_HPP_
+#ifndef PANTHER_UTILS_ROS_UTILS_HPP_
+#define PANTHER_UTILS_ROS_UTILS_HPP_
 
 #include <chrono>
 
-#include <tf2_ros/buffer.h>
-
-#include <geometry_msgs/msg/pose_stamped.hpp>
 #include <std_msgs/msg/header.hpp>
 
 namespace panther_utils::ros
@@ -118,31 +115,6 @@ std::string AddNamespaceToFrameID(const std::string & frame_id, const std::strin
   return tf_prefix + frame_id;
 }
 
-geometry_msgs::msg::PoseStamped TransformPose(
-  const tf2_ros::Buffer::SharedPtr & tf2_buffer, const geometry_msgs::msg::PoseStamped & pose,
-  const std::string & target_frame, double timeout_s = 0.0)
-{
-  geometry_msgs::msg::PoseStamped transformed_pose;
-
-  if (pose.header.frame_id.empty() || target_frame.empty()) {
-    throw std::runtime_error(
-      "Pose or target frame is empty, pose frame: \"" + pose.header.frame_id +
-      "\", target frame: \"" + target_frame + "\"");
-  }
-
-  if (!tf2_buffer->canTransform(
-        pose.header.frame_id, target_frame, pose.header.stamp,
-        rclcpp::Duration::from_seconds(timeout_s))) {
-    throw std::runtime_error(
-      "Cannot transform " + pose.header.frame_id + " to " + target_frame + " at time " +
-      std::to_string(pose.header.stamp.sec) + "." + std::to_string(pose.header.stamp.nanosec));
-  }
-
-  tf2_buffer->transform(pose, transformed_pose, target_frame);
-
-  return transformed_pose;
-}
-
 }  // namespace panther_utils::ros
 
-#endif  // PANTHER_UTILS__ROS_UTILS_HPP_
+#endif  // PANTHER_UTILS_ROS_UTILS_HPP_
